move state component
