+++ extensions/homeassistant/src/binarysensors.tsx
@@ -1,4 +1,4 @@
-import { StatesList } from "./components";
+import { StatesList } from "./components/states";
 
 export default function main() {
   return <StatesList domain="binary_sensor" />;
